scripts/requirements.txt
pdfplumber
psycopg2-binary
python-dotenv
